move common methods
